Add calculator-stream service api and impl
