+++ calculator-stream-impl/src/main/resources/application.conf
@@ -1,0 +1,1 @@
+play.modules.enabled += calculator.stream.impl.CalculatorStreamModule
